ai comments, refactor to XAudioUtils
